[#OS-234], routing issue fixed for no trips and bids, added bank transction id for advance paid

Approved-by: Saurabh Dhillon <saurabh.dhillon@delhivery.com>

+++ app/src/main/java/com/delhivery/axle/ui/biddetails/BidDetailsUserBidState.kt
@@ -23,7 +23,7 @@ data class BidDetailsUserBidState_PlaceBidFirst(val bidsCount: Int = 0) : BidDet
 data class BidDetailsUserBidState_PlaceBid(
   val bidsCount: Int,
   val bids: List<TransactionBid>,
-  val lowestBid: Int
+  val lowestBid: Double
 ) : BidDetailsUserBidState(R.layout.view_bid_details_place_bid)
 
 /**
@@ -33,7 +33,7 @@ data class BidDetailsUserBidState_EditBid(
   val bidsCount: Int,
   val bids: List<TransactionBid>,
   val userBid: TransactionBid,
-  val lowestBid: Int
+  val lowestBid: Double
 ) : BidDetailsUserBidState(R.layout.view_bid_details_edit_bid)
 
 /**
